Add JNeT demo code
There are numerous issues with this, since it's based on an outdated
example using an older version of JNeT. I've had to make some changes
to accommodate the different hardware (original example was for the
F100) and newer version of JNeT and with this one especially, I have no
idea what I'm doing.
As a result, there's a lot of weird behavior that could be caused by
the hardware, or the code. Just perfect.

The first areas to investigate are the truncate calls that I've left out
where the now unused length variables are. The Ethernet frame should be
truncated with these, but thanks to a reorganization of JNeT with the
owning slice, I don't really know how to best do that yet.