Remove ITM tracing and switch to serial
This is unfortunate, but on my device, it exhibited very strange
failures. After some repeated writing to it, it kind of froze and
stopped outputting in the middle of some data, only to continue later
with new data. It can be "fixed" by putting delays between print calls,
but at that point it's more efficient to just downgrade to serial.

I have no idea if it's a hardware or software problem. In any case,
it shouldn't be my code. I ran several ITM examples from other sources
and they have the same issue.