Do software reset with SPI instead of RESET pin
The enc28j60 driver resets it by setting the RST pin to low and
immediately to high again. According to the data sheet, it needs to be
at low for at least 400 ns, which is not guaranteed with this method.
As a result the ENC28J60 is not properly reset, which can lead to
already received packets being returned from the internal buffer again
and weird behavior in general.

One solution is to introduce a delay before setting RESET to high again,
but doing an SPI reset works just as well and requires a wire less.